BugFix: RowSlider does not affect falling tetromino. + Bugfix: Tetromino falls to the correct location even after DestroyRow function.

+++ Assets/Scripts/Map/Mino.cs
@@ -8,7 +8,7 @@ public class Mino : MonoBehaviour
 
 
     public float originalPosition;
-    public int slideDestination;
+    public float slideDestination;
 
     public Vector3Int localPosition;
